Break apart the transaction details styles
