+++ src/components/common/ContactSearchResults.js
@@ -1,12 +1,13 @@
 // @flow
 
 import React, { Component } from 'react'
-import { FlatList, Image, TouchableHighlight, View } from 'react-native'
+import { FlatList, Image, StyleSheet, TouchableHighlight, View } from 'react-native'
 
 import ContactImage from '../../assets/images/contact.png'
 import FormattedText from '../../modules/UI/components/FormattedText/FormattedText.ui.js'
-import styles from '../../styles/scenes/TransactionDetailsStyle.js'
+import { THEME } from '../../theme/variables/airbitz.js'
 import { type GuiContact } from '../../types/types.js'
+import { scale } from '../../util/scaling.js'
 
 type Props = {
   contacts: Array<GuiContact>,
@@ -54,11 +55,7 @@ export class ContactSearchResults extends Component<Props> {
     const fullName = familyName ? `${givenName} ${familyName}` : givenName
     return (
       <View style={styles.singleContactWrap}>
-        <TouchableHighlight
-          style={styles.singleContact}
-          onPress={() => this.props.onSelectPayee(fullName, thumbnailPath)}
-          underlayColor={styles.underlayColor.color}
-        >
+        <TouchableHighlight style={styles.singleContact} onPress={() => this.props.onSelectPayee(fullName, thumbnailPath)} underlayColor={THEME.COLORS.GRAY_4}>
           <View style={styles.contactInfoWrap}>
             <View style={styles.contactLeft}>
               <View style={styles.contactLogo}>
@@ -78,3 +75,52 @@ export class ContactSearchResults extends Component<Props> {
     )
   }
 }
+
+const rawStyles = {
+  resultList: {
+    backgroundColor: THEME.COLORS.WHITE,
+    borderTopColor: THEME.COLORS.GRAY_3,
+    borderTopWidth: 1,
+    flex: 1
+  },
+  singleContact: {
+    height: scale(60),
+    borderBottomWidth: 1,
+    borderBottomColor: THEME.COLORS.TRANSACTION_DETAILS_GREY_2,
+    padding: scale(10),
+    paddingRight: scale(15),
+    paddingLeft: scale(15)
+  },
+  singleContactWrap: {
+    flexDirection: 'column',
+    flex: 1
+  },
+  contactInfoWrap: {
+    flexDirection: 'row',
+    height: scale(40),
+    flex: 1,
+    justifyContent: 'space-between'
+  },
+  contactLeft: {
+    flexDirection: 'row'
+  },
+  contactLogo: {
+    width: scale(40),
+    height: scale(40),
+    marginRight: scale(10)
+  },
+  contactLeftTextWrap: {
+    justifyContent: 'center'
+  },
+  contactName: {
+    fontSize: scale(16),
+    color: THEME.COLORS.TRANSACTION_DETAILS_GREY_3,
+    textAlignVertical: 'center'
+  },
+  contactThumbnail: {
+    height: scale(40),
+    width: scale(40),
+    borderRadius: scale(20)
+  }
+}
+const styles: typeof rawStyles = StyleSheet.create(rawStyles)
