+++ src/components/scenes/TransactionDetailsScene.js
@@ -19,8 +19,8 @@ import { getDisplayDenomination, getPlugins, getSettings } from '../../modules/S
 import { PrimaryButton2 } from '../../modules/UI/components/Buttons/PrimaryButton2.ui.js'
 import FormattedText from '../../modules/UI/components/FormattedText/FormattedText.ui.js'
 import { convertCurrencyFromExchangeRates, convertNativeToExchangeRateDenomination, getSelectedWallet, getWallet } from '../../modules/UI/selectors.js'
-import { iconSize } from '../../styles/scenes/TransactionDetailsStyle.js'
 import { type ThemeProps, cacheStyles, withTheme } from '../../theme/ThemeContext.js'
+import { THEME } from '../../theme/variables/airbitz.js'
 import { type Dispatch, type State as ReduxState } from '../../types/reduxTypes.js'
 import type { GuiContact, GuiWallet } from '../../types/types.js'
 import { scale } from '../../util/scaling.js'
@@ -445,7 +445,7 @@ export class TransactionDetailsComponent extends Component<Props, State> {
                   {thumbnailPath ? (
                     <Image style={styles.tileThumbnail} source={{ uri: thumbnailPath }} />
                   ) : (
-                    <IonIcon style={styles.tileAvatarIcon} name="ios-contact" size={iconSize.avatar} />
+                    <IonIcon style={styles.tileAvatarIcon} name="ios-contact" size={THEME.rem(2)} />
                   )}
                   <FormattedText style={styles.tileTextBottom}>{personName}</FormattedText>
                 </View>
