+++ src/constants/IconConstants.js
@@ -56,7 +56,6 @@ export const ARROW_DOWN_THICK = 'arrow-down-thick'
 export const ARROW_DOWN_BOLD = 'arrow-down-bold'
 export const ARROW_DOWN = 'arrow-down'
 export const PICKAXE = 'pickaxe'
-export const TELESCOPE = 'telescope'
 export const SERVER = 'server'
 export const CLOSE_CIRCLE_ICON = 'close-circle-outline'
 export const CHECK_CIRCLE_ICON = 'check-circle-outline'
